Move question type-specific response validation to QuestionDetails class
