+++ src/main/java/teammates/common/datatransfer/questions/FeedbackNumericalScaleQuestionDetails.java
@@ -42,6 +42,35 @@ public class FeedbackNumericalScaleQuestionDetails extends FeedbackQuestionDetai
     }
 
     @Override
+    public List<String> validateResponsesDetails(List<FeedbackResponseDetails> responses) {
+        List<String> errors = new ArrayList<>();
+
+        for (FeedbackResponseDetails response : responses) {
+            FeedbackNumericalScaleResponseDetails details = (FeedbackNumericalScaleResponseDetails) response;
+
+            // out of range
+            boolean isAnswerOutOfRange = details.getAnswer() < minScale || details.getAnswer() > maxScale;
+            if (isAnswerOutOfRange) {
+                errors.add(details.getAnswerString() + Const.FeedbackQuestion.NUMSCALE_ERROR_OUT_OF_RANGE
+                        + "(min=" + minScale + ", max=" + maxScale + ")");
+            }
+
+            // when the answer is within range but not one of the possible values
+            double remainder = Double.valueOf(String.format("%.5f", (details.getAnswer() - minScale) % step));
+            boolean isAnswerNotAPossibleValueWithinRange = remainder != 0.0 && !isAnswerOutOfRange;
+
+            if (isAnswerNotAPossibleValueWithinRange) {
+                double nextPossibleValueLessThanCurrent = details.getAnswer() - remainder;
+                double nextPossibleValueGreaterThanCurrent = nextPossibleValueLessThanCurrent + step;
+                errors.add("Please enter a valid value. The two nearest valid values are "
+                        + nextPossibleValueLessThanCurrent + " and " + nextPossibleValueGreaterThanCurrent + ".");
+            }
+        }
+
+        return errors;
+    }
+
+    @Override
     public boolean isFeedbackParticipantCommentsOnResponsesAllowed() {
         return false;
     }
