+++ src/main/java/teammates/common/datatransfer/questions/FeedbackRankRecipientsResponseDetails.java
@@ -1,10 +1,5 @@
 package teammates.common.datatransfer.questions;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import teammates.common.datatransfer.attributes.FeedbackQuestionAttributes;
-
 public class FeedbackRankRecipientsResponseDetails extends FeedbackResponseDetails {
     private int answer;
 
@@ -15,11 +10,6 @@ public class FeedbackRankRecipientsResponseDetails extends FeedbackResponseDetai
     @Override
     public String getAnswerString() {
         return Integer.toString(answer);
-    }
-
-    @Override
-    public List<String> validateResponseDetails(FeedbackQuestionAttributes correspondingQuestion) {
-        return new ArrayList<>();
     }
 
     public int getAnswer() {
